gui: rework localization support
Drop string localization for the text in the GUI and switch to
hard-coded localized text instead. Unfortunately there's no way to
localize strings on mods without having the localization file installed.
To work around this, add the localized text directly.

Also try to auto-detect the language and set the text accordingly. As a
fallback, we also now have buttons on the menu to change the language at
will.